New feature: addition of TextToSpeech for the Android platform

+++ AudioJDR/Model/ISpeechSynthesizer.cs
@@ -61,8 +61,8 @@ namespace Model
         /// Sets the rate at which the voice speaks
         /// Adjust the speed of the speech output
         /// </summary>
-        /// <param name="voiceRate">The rate of the voice (between -10 and 10)</param>
-        void SetRate(int voiceRate);
+        /// <param name="voiceRate">The rate of the voice (a float between 0.5f and 2.0f)</param>
+        void SetRate(float voiceRate);
 
         /// <summary>
         /// Releases any resources used by the speech synthesizer
